Added resevations, few fixes and redos

+++ app/src/Controller/CategoryController.php
@@ -1,5 +1,4 @@
 <?php
-
 /**
  * Category controller.
  */
@@ -9,66 +8,64 @@ namespace App\Controller;
 use App\Entity\Category;
 use App\Form\Type\CategoryType;
 use App\Service\CategoryServiceInterface;
+use App\Service\TaskServiceInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\HttpKernel\Attribute\MapQueryParameter;
-use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\Security\Http\Attribute\IsGranted;
 use Symfony\Contracts\Translation\TranslatorInterface;
+use Symfony\Component\Form\Extension\Core\Type\FormType;
 
 /**
  * Class CategoryController.
  */
-#[Route('/category')]
+#[\Symfony\Component\Routing\Attribute\Route('/category')]
 class CategoryController extends AbstractController
 {
     /**
      * Constructor.
      *
      * @param CategoryServiceInterface $categoryService Category service
+     * @param TaskServiceInterface     $taskService     Task service
      * @param TranslatorInterface      $translator      Translator
      */
-    public function __construct(private readonly CategoryServiceInterface $categoryService,  private readonly TranslatorInterface $translator)
+    public function __construct(private readonly CategoryServiceInterface $categoryService, private readonly TaskServiceInterface $taskService, private readonly TranslatorInterface $translator)
     {
     }
 
     /**
      * Index action.
      *
-     * @param int $page Page number
+     * @param Request $request The HTTP request
      *
      * @return Response HTTP response
      */
-    #[Route(
-        name: 'category_index',
-        methods: ['GET']
-    )]
-    public function index(#[MapQueryParameter] int $page = 1): Response
+    #[\Symfony\Component\Routing\Attribute\Route(name: 'category_index', methods: 'GET')]
+    public function index(Request $request): Response
     {
+        $page = $request->query->getInt('page', 1);
         $pagination = $this->categoryService->getPaginatedList($page);
 
         return $this->render('category/index.html.twig', ['pagination' => $pagination]);
     }
 
     /**
-     * View action.
+     * Show action.
      *
-     * @param Category $category Category entity
+     * @param Category $category Category
      *
      * @return Response HTTP response
      */
-    #[Route(
+    #[\Symfony\Component\Routing\Attribute\Route(
         '/{id}',
-        name: 'category_view',
+        name: 'category_show',
         requirements: ['id' => '[1-9]\d*'],
-        methods: ['GET']
+        methods: 'GET'
     )]
-    public function view(Category $category): Response
+    #[IsGranted('VIEW_CATEGORY', subject: 'category')]
+    public function show(Category $category): Response
     {
-        return $this->render(
-            'category/view.html.twig',
-            ['category' => $category]
-        );
+        return $this->render('category/show.html.twig', ['category' => $category]);
     }
 
     /**
@@ -78,11 +75,12 @@ class CategoryController extends AbstractController
      *
      * @return Response HTTP response
      */
-    #[Route(
+    #[\Symfony\Component\Routing\Attribute\Route(
         '/create',
         name: 'category_create',
-        methods: ['GET', 'POST']
+        methods: 'GET|POST',
     )]
+    #[IsGranted('ROLE_ADMIN')]
     public function create(Request $request): Response
     {
         $category = new Category();
@@ -114,12 +112,8 @@ class CategoryController extends AbstractController
      *
      * @return Response HTTP response
      */
-    #[Route(
-        '/{id}/edit',
-        name: 'category_edit',
-        requirements: ['id' => '[1-9]\d*'],
-        methods: ['GET', 'PUT']
-    )]
+    #[\Symfony\Component\Routing\Attribute\Route('/{id}/edit', name: 'category_edit', requirements: ['id' => '[1-9]\d*'], methods: 'GET|PUT')]
+    #[IsGranted('EDIT_CATEGORY', subject: 'category')]
     public function edit(Request $request, Category $category): Response
     {
         $form = $this->createForm(
@@ -137,7 +131,7 @@ class CategoryController extends AbstractController
 
             $this->addFlash(
                 'success',
-                $this->translator->trans('message.edited_successfully')
+                $this->translator->trans('message.updated_successfully')
             );
 
             return $this->redirectToRoute('category_index');
@@ -160,18 +154,27 @@ class CategoryController extends AbstractController
      *
      * @return Response HTTP response
      */
-    #[Route(
-        '/{id}/delete',
-        name: 'category_delete',
-        requirements: ['id' => '[1-9]\d*'],
-        methods: ['GET', 'DELETE']
-    )]
+    #[\Symfony\Component\Routing\Attribute\Route('/{id}/delete', name: 'category_delete', requirements: ['id' => '[1-9]\d*'], methods: 'GET|DELETE')]
+    #[IsGranted('DELETE_CATEGORY', subject: 'category')]
     public function delete(Request $request, Category $category): Response
     {
-        $form = $this->createForm(FormType::class, $category, [
-            'method' => 'DELETE',
-            'action' => $this->generateUrl('category_delete', ['id' => $category->getId()]),
-        ]);
+        if (!$this->categoryService->canBeDeleted($category)) {
+            $this->addFlash(
+                'warning',
+                $this->translator->trans('message.category_contains_tasks')
+            );
+
+            return $this->redirectToRoute('category_index');
+        }
+
+        $form = $this->createForm(
+            FormType::class,
+            $category,
+            [
+                'method' => 'DELETE',
+                'action' => $this->generateUrl('category_delete', ['id' => $category->getId()]),
+            ]
+        );
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
@@ -192,5 +195,29 @@ class CategoryController extends AbstractController
                 'category' => $category,
             ]
         );
+    }
+
+    /**
+     * Show tasks action.
+     *
+     * @param Category $category Category
+     *
+     * @return Response HTTP response
+     */
+    #[\Symfony\Component\Routing\Attribute\Route(
+        '/{id}/tasks',
+        name: 'category_show_tasks',
+        requirements: ['id' => '[1-9]\d*'],
+        methods: 'GET'
+    )]
+    #[IsGranted('VIEW_CATEGORY', subject: 'category')]
+    public function showTasks(Category $category): Response
+    {
+        $tasks = $this->taskService->getTasksByCategory($category);
+
+        return $this->render('category/show_tasks.html.twig', [
+            'category' => $category,
+            'tasks' => $tasks,
+        ]);
     }
 }
